+++ utopia/scripts/make_basic.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
+set -e
+set -o pipefail
 
 today=$(date)
 printf "%s\n" "$today"
@@ -11,35 +13,39 @@ else
 	N_THREADS=$1
 fi
 
+prefix=$2
+
 _basic_build(){
 
 	touch make_basic.log
 
-	cmake .. -DUTOPIA_ENABLE_BLAS=ON -DUTOPIA_ENABLE_TRILINOS=OFF -DUTOPIA_ENABLE_PETSC=OFF -DUTOPIA_ENABLE_EXAMPLES=ON -DUTOPIA_ENABLE_TESTS=ON -DCMAKE_INSTALL_PREFIX=$2 | tee make_basic.log
+	cmake .. -DUTOPIA_ENABLE_BLAS=ON -DUTOPIA_ENABLE_TRILINOS=OFF -DUTOPIA_ENABLE_PETSC=OFF -DUTOPIA_ENABLE_EXAMPLES=ON -DUTOPIA_ENABLE_TESTS=ON -DCMAKE_INSTALL_PREFIX=$prefix | tee make_basic.log
 
 
-	make -j$N_THREADS complete | tee -a make_basic.log
-	./utopia_bench | tee -a make_basic.log
-	./utopia_test | tee -a make_basic.log
-	make install | tee -a make_basic.log
-	make -j$N_THREADS test_install | tee -a make_basic.log
+	read -p "Continue ? y/n" -n 1 -r
+	echo "\n"    # (optional) move to a new line
+	if [[ $REPLY =~ ^[Yy]$ ]]
+	then
+	    # do dangerous stuff
+	    make -j$N_THREADS complete | tee -a make_basic.log
+		./utopia_bench | tee -a make_basic.log
+		./utopia_test | tee -a make_basic.log
+		make install | tee -a make_basic.log
+		make -j$N_THREADS test_install | tee -a make_basic.log
+	fi
 }
-
-
-jobs=$1
-prefix=$2
 
 if [[ -d build_basic ]]
 then
 	cd build_basic
 	rm -rf *
-	_basic_build $jobs $prefix
-fi
-
-if [[ ! -d build_basic ]]
-then
-	mkdir build_basic
-	cd build_basic
-	rm -rf *
-	_basic_build $jobs $prefix
+	_basic_build
+else
+	if [[ ! -d build_basic ]]
+	then
+		mkdir build_basic
+		cd build_basic
+		rm -rf *
+		_basic_build
+	fi
 fi
